requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
